+++ api/models/Role.spec.js
@@ -6,8 +6,9 @@ const {
   roleDefaults,
   PermissionTypes,
 } = require('librechat-data-provider');
-const { getRoleByName, updateAccessPermissions, initializeRoles } = require('~/models/Role');
+const { getRoleByName, updateAccessPermissions } = require('~/models/Role');
 const getLogStores = require('~/cache/getLogStores');
+const { initializeRoles } = require('~/models');
 
 const Role = require('~/db/models').Role;
